+++ app/models/document_revision.py
@@ -15,8 +15,8 @@ class DocumentRevision(Base):
     __tablename__ = "documents_revisions"
     __table_args__ = (
         UniqueConstraint(
-            "document_id", "revision",
-            name="uq_documents_revisions_document_id_revision"
+            "document_id", "revision_number",
+            name="uq_documents_revisions_document_id_revision_number"
         ),
         {"sqlite_autoincrement": True},
     )
@@ -49,27 +49,21 @@ class DocumentRevision(Base):
         default=lambda: int(time.time())
     )
 
-    revision = Column(
+    revision_number = Column(
         Integer,
         nullable=False,
         index=True,
     )
 
-    filename = Column(
+    uuid = Column(
         String(256),
-        nullable=False,
-        index=True
+        index=True,
+        nullable=False
     )
 
     filesize = Column(
         Integer,
         nullable=False,
-        index=True
-    )
-
-    mimetype = Column(
-        String(256),
-        nullable=True,
         index=True
     )
 
@@ -79,9 +73,13 @@ class DocumentRevision(Base):
         index=True
     )
 
+    # NOTE: Join users because revisions themselves are not cached;
+    # they should be cached as part of documents and contain users.
     revision_user = relationship(
         "User",
-        back_populates="user_revisions"
+        back_populates="user_revisions",
+        uselist=False,
+        lazy="joined"
     )
 
     revision_document = relationship(
@@ -90,13 +88,24 @@ class DocumentRevision(Base):
     )
 
     def __init__(
-            self, user_id: int, document_id: int, revision: int,
-            filename: str, filesize: int, checksum: str,
-            mimetype: str = None):
+            self, user_id: int, document_id: int, revision_number: int,
+            uuid: str, filesize: int, checksum: str):
         self.user_id = user_id
         self.document_id = document_id
-        self.revision = revision
-        self.filename = filename
+        self.revision_number = revision_number
+        self.uuid = uuid
         self.filesize = filesize
-        self.mimetype = mimetype
         self.checksum = checksum
+
+    async def to_dict(self) -> dict:
+        """Returns a dictionary representation of the revision."""
+        return {
+            "id": self.id,
+            "user": await self.revision_user.to_dict(),
+            "document_id": self.document_id,
+            "created_date": self.created_date,
+            "revision_number": self.revision_number,
+            "uuid": self.uuid,
+            "filesize": self.filesize,
+            "checksum": self.checksum,
+        }